Added support for permanent audio segments and memory cache size limiting.

+++ rad_compositor/src/composition.rs
@@ -7,7 +7,7 @@ use crate::source::{BaseSource, Source, TFrameIdx};
 pub type TWrappedCompositionState = Arc<RwLock<CompositionState>>;
 
 static COMPOSITION_ID_TO_ALLOCATE: Mutex<u16> = Mutex::new(0);
-pub struct SrcCompositionData { pub frame_offset: isize, pub amplification: f32 }
+pub struct SrcCompositionData { pub frame_offset: i64, pub amplification: f32 }
 pub struct CompositionSrc { pub src: Source, pub composition_data: SrcCompositionData }
 
 pub fn convert_sample_rates(sample_rate_a: TFrameIdx, rate_a: TFrameIdx, sample_rate_b: TFrameIdx) -> TFrameIdx {
@@ -48,14 +48,14 @@ impl CompositionState {
 		self.sources.push(CompositionSrc {
 			composition_data: SrcCompositionData { 
 				amplification: 1.0,
-				frame_offset: (self.get_time_sec() * src.sample_rate() as f64) as isize
+				frame_offset: (self.get_time_sec() * src.sample_rate() as f64) as i64
 			},
 			src,
 		});
 
 	}
 
-	pub fn push_src_offset(&mut self, src: Source, frame_offset: isize) {
+	pub fn push_src_offset(&mut self, src: Source, frame_offset: i64) {
 		self.sources.push(CompositionSrc {
 			composition_data: SrcCompositionData { 
 				amplification: 1.0,
